WIP- adds embed code to useEffect in <SupersetDashboard/>; 404ing
Code is successfully mounting an iFrame to the DOM, but it contains
only a 404 JSON response payload.  Additionally, there is an error
regarding `Buffer` not being defined.  I believe these *may* be two
separate and unrelated issues; my suspicion is:

- The Buffer issue is because the embedded SDK is not intended for
  browser usage, and can hopefully be resolved with a polyfill for
  the Buffer object
- While it is possible the 404 is due to the Buffer issue, I think
  instead it might be due to missing configuration for the Flask
  process to allow/enable embedded dashboards.

These will all be targeted in subsequent commits.  Once a working
POC of the dashboard alone is complete, proper plumbing for
fetching tokens will be implemented before moving on to the
bidirectional interactions.