Styles dashboard iframe to fill the space appropriately
This apparently is working with the hardcoded credentials and guest
pass even a day later.  That said, it might still be worthwhile to
put some basic automatic auth in place for others that may be using
this repo as an example (with messaging explaining it is not to be
used this way in production).